+++ dwds/test/chrome_proxy_service_test.dart
@@ -652,6 +652,23 @@ void main() {
           expect(sixth.valueAsString, '5');
         });
 
+        test(
+            'Lists with null count and offset greater than 0 are '
+            'truncated from offset to end of list', () async {
+          final list = await createList();
+          final inst = await service.getObject(
+            isolate.id!,
+            list.objectId!,
+            count: null,
+            offset: 1000,
+          ) as Instance;
+          expect(inst.length, 1001);
+          expect(inst.offset, 1000);
+          expect(inst.count, null);
+          final only = inst.elements![0] as InstanceRef;
+          expect(only.valueAsString, '5');
+        });
+
         test('Lists with offset/count are truncated', () async {
           final list = await createList();
           final inst = await service.getObject(
@@ -702,6 +719,24 @@ void main() {
           final sixth = inst.associations![5];
           expect(sixth.key.valueAsString, '5');
           expect(sixth.value.valueAsString, '995');
+        });
+
+        test(
+            'Maps with null count and offset greater than 0 are '
+            'truncated from offset to end of map', () async {
+          final list = await createMap();
+          final inst = await service.getObject(
+            isolate.id!,
+            list.objectId!,
+            count: null,
+            offset: 1000,
+          ) as Instance;
+          expect(inst.length, 1001);
+          expect(inst.offset, 1000);
+          expect(inst.count, null);
+          final only = inst.associations![0];
+          expect(only.key.valueAsString, '1000');
+          expect(only.value.valueAsString, '0');
         });
 
         test('Maps with null count are not truncated', () async {